Extract reusable item stack popup
